➕ Component for admin drawer
